README.md - adds guide to fix lack of spacing on vscode
I don't know if this happens to anyone else but if I don't explicitly set my
"terminal.integrated.letterSpacing" to 1 the font will render like there is a
zero (or a -1) on the spacing inside the VSCode terminal window.

To fix it we just need to set the property "terminal.integrated.letterSpacing"
to 1 and font will render properly.

This commit adds a quick guide (with screenshots) about how to fix the issue
described above.